LayoutTests/fast/page-color-sampling/color-sampling-ignores-backdrop-filters.html

<!DOCTYPE html> <!-- webkit-test-runner [ ContentInsetBackgroundFillEnabled=true pageTopColorSamplingEnabled=true useFlexibleViewport=true obscuredInset.top=75 ] -->
<html>
<head>
    <meta charset="utf-8">
    <meta name="viewport" content="width=device-width, initial-scale=1">
    <style>
        body {
            font-family: system-ui;
        }

        header {
            font-size: 40px;
            line-height: 100px;
            text-align: center;
            position: fixed;
            top: 0;
            left: 0;
            width: 100%;
            height: 100px;
            backdrop-filter: blur(40px);
            color: white;
        }

        h1 {
            margin-top: 100px;
            text-align: center;
        }

        .gradient {
            margin: 1em 0 500px 0;
            width: 100%;
            height: 1000px;
            background: linear-gradient(to bottom, red 0%, green 50%, blue 100%);
        }

        footer {
            position: fixed;
            bottom: 0;
            left: 0;
            width: 100%;
            height: 100px;
            color: white;
            font-size: 20px;
            line-height: 100px;
            background: lightgray;
            text-align: center;
        }
    </style>
    <script src="../../resources/js-test.js"></script>
    <script src="../../resources/ui-helper.js"></script>
    <script>
    jsTestIsAsync = true;

    addEventListener("load", async () => {
        await UIHelper.setObscuredInsets(75, 0, 75, 0);
        await UIHelper.ensurePresentationUpdate();
        sampledColors = [];
        for (i = 0; i < 5; ++i) {
            sampledColors.push(await UIHelper.fixedContainerEdgeColors());
            scrollBy(0, 100);
            await UIHelper.ensurePresentationUpdate();
        }

        for (colors of sampledColors) {
            shouldBeEqualToString("colors.top", "multiple");
            shouldBeNull("colors.left");
            shouldBeNull("colors.right");
            shouldBeEqualToString("colors.bottom", "rgb(211, 211, 211)");
        }

        finishJSTest();
    });
    </script>
</head>
<body>
<header>Headline</header>
<h1>Headline</h1>
<div class="gradient"></div>
<pre id="console"></pre>
<pre id="description"></pre>
<footer>Bottom bar</footer>
</body>
</html>
