LayoutTests/fast/page-color-sampling/color-sampling-ignores-backdrop-filters-expected.txt

Headline
Headline

PASS colors.top is null
PASS colors.left is null
PASS colors.right is null
PASS colors.bottom is "rgb(211, 211, 211)"
PASS colors.top is null
PASS colors.left is null
PASS colors.right is null
PASS colors.bottom is "rgb(211, 211, 211)"
PASS colors.top is null
PASS colors.left is null
PASS colors.right is null
PASS colors.bottom is "rgb(211, 211, 211)"
PASS colors.top is null
PASS colors.left is null
PASS colors.right is null
PASS colors.bottom is "rgb(211, 211, 211)"
PASS colors.top is null
PASS colors.left is null
PASS colors.right is null
PASS colors.bottom is "rgb(211, 211, 211)"
PASS successfullyParsed is true

TEST COMPLETE
Bottom bar
